+++ challenge_model/src/main/java/com/amedigital/challenge_model/repositories/ProdutoRepository.kt
@@ -6,7 +6,7 @@ import com.amedigital.challenge_model.api.Resource
 interface ProdutoRepository {
     suspend fun getProdutos(): Resource<List<Produto>>
     suspend fun getMaisVendidos(): Resource<List<Produto>>
-    suspend fun getByCategoria(categoriaId: Long): Resource<List<Produto>>
-    suspend fun getProduto(produtoId: Long): Resource<Produto>
-    suspend fun reservarProduto(produtoId: Long): Resource<String>
+    suspend fun getByCategoria(categoriaId: Int): Resource<List<Produto>>
+    suspend fun getProduto(produtoId: Int): Resource<Produto>
+    suspend fun reservarProduto(produtoId: Int): Resource<String>
 }
